refactor network context
